fix: get oracle before 24 hours
